event_booking_backend/requirements.txt

fastapi==0.109.0
uvicorn==0.24.0
motor==3.3.1
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.0
bcrypt==4.1.2
pymongo==4.6.1
python-jose==3.3.0
email-validator==2.1.0
PyJWT==2.8.0
python-dateutil==2.9.0